Moved static getAnalyzer to HighlightUtils
